refactored the service to use resource issuer for security
