+++ application.yml
@@ -1,0 +1,29 @@
+spring:
+  application:
+    name: onboarding-service
+  datasource:
+    url: jdbc:postgresql://localhost:5432/mydatabase
+    username: myuser
+    password: mypassword
+    driver-class-name: org.postgresql.Driver
+  jpa:
+    hibernate:
+      ddl-auto: update
+    show-sql: true
+  kafka:
+    bootstrap-servers: localhost:9092
+  flyway:
+    baseline-on-migrate: true
+
+  security:
+    oauth2:
+      resourceserver:
+        jwt:
+          issuer-uri: http://localhost:8080
+server:
+  port: 8083
+
+logging:
+  level:
+    root: INFO
+    org.apache.kafka: DEBUG
